Agent/skills/ppt-generator/requirements.txt
python-pptx>=1.0.2
openai>=1.40.0
python-dotenv>=1.0.1
